fix MAX_SIZE_T
Take into account the fact that we multiply arguments to malloc()
and realloc() by sizeof(wchar_t). So, the total buffer capacity
reduces in sizeof(wchar_t)-times if we use wide characters,
and in the worst case the number of chars that can be read from file
reduces in 5 times more (we estimate the upper bound by file size
in bytes, and maximum possible number of bytes per char in UTF-8 is
6 bytes, which gives 5 items unused).

Hmm... So, if we have 16-bit size_t and 4-byte wchar_t, this becomes

    65536 / 4 / 5

this gives ~ 3276 chars maximum that can be read. Not too much...