use DB_FILE for locking and for saving cursor
Lock file was not created previously, because it was created in
the same directory as wanted file, and this directory could be
owned by, e.g., root. So, a mechanism was
invented not to create lock file in such cases. But now locking
always may be done because separate file is used. Moreover,
locking is necessary even when file is read-only --- for consistency
in handling saving/restoring cursor.